Folding: reorganize examples + add explanation for the user
